# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Streetly API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /riders/apply:
    post:
      operationId: applyAsRider
      tags: [riders]
      summary: Apply to become a delivery rider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
  /riders/{riderId}/status:
    patch:
      operationId: setRiderOnlineStatus
      tags: [riders]
      summary: Go online or offline
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderOnlineStatusInput"
      responses:
        "200":
          description: Updated rider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
  /riders/{riderId}/location:
    post:
      operationId: updateRiderLocation
      tags: [riders]
      summary: Push a live GPS location update
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderLocationInput"
      responses:
        "200":
          description: Updated rider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
  /riders/nearby:
    get:
      operationId: getNearbyRiders
      tags: [riders]
      summary: Get online riders near a point
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
        - name: radiusKm
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Nearby riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyRider"
  /riders/available:
    get:
      operationId: getAvailableRidersDirectory
      tags: [riders]
      summary: List all currently online/approved riders for the public rider directory
      parameters:
        - name: lat
          in: query
          required: false
          schema:
            type: number
        - name: lon
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Available riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiderDirectoryEntry"
  /riders/{riderId}/orders:
    get:
      operationId: getRiderOrders
      tags: [riders]
      summary: Get a rider's active order and available nearby requests
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rider order queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderOrderQueue"

  # Deliveries
...
components:
  schemas:
...
    Rider:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        fullName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        idType:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        passportObjectPath:
          type: ["string", "null"]
        ninSlipObjectPath:
          type: ["string", "null"]
        isOnline:
          type: boolean
        currentLatitude:
          type: ["number", "null"]
        currentLongitude:
          type: ["number", "null"]
        lastLocationAt:
          type: ["string", "null"]
        totalDeliveries:
          type: integer
        createdAt:
          type: string
    RiderOnlineStatusInput:
      type: object
      required: [isOnline]
      properties:
        isOnline:
          type: boolean
    RiderLocationInput:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
        longitude:
          type: number
    RiderApplicationInput:
      type: object
      required: [fullName, phone, vehicleType, dateOfBirth, address, passportObjectPath, ninSlipObjectPath]
      properties:
        fullName:
          type: string
        phone:
          type: string
        vehicleType:
          type: string
        idType:
          type: string
        idNumber:
          type: string
        dateOfBirth:
          type: string
          description: Date of birth (YYYY-MM-DD).
        address:
          type: string
          description: Full residential address.
        passportObjectPath:
          type: string
          description: Object storage path for the uploaded passport photo/document.
        ninSlipObjectPath:
          type: string
          description: Object storage path for the uploaded NIN slip document.
    NearbyRider:
      type: object
      required: [id, currentLatitude, currentLongitude, distanceKm]
      properties:
        id:
          type: integer
        fullName:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        currentLatitude:
          type: number
        currentLongitude:
          type: number
        lastLocationAt:
          type: ["string", "null"]
        distanceKm:
          type: number
    RiderDirectoryEntry:
      type: object
      required: [id, totalDeliveries, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        currentLatitude:
          type: ["number", "null"]
        currentLongitude:
          type: ["number", "null"]
        lastLocationAt:
          type: ["string", "null"]
        distanceKm:
          type: ["number", "null"]
        totalDeliveries:
          type: integer
        createdAt:
          type: string
    RiderOrderQueue:
      type: object
      required: [active, available, completedCount]
      properties:
        active:
          type: array
          items:
            $ref: "#/components/schemas/DeliveryOrder"
        available:
          type: array
          items:
            $ref: "#/components/schemas/DeliveryOrder"
        completedCount:
          type: integer

    # Deliveries
    DeliveryOrder:
      type: object
      required: [id, businessId, customerName, customerPhone, deliveryAddress, status, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        customerUserId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        deliveryLatitude:
          type: ["number", "null"]
        deliveryLongitude:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [requested, accepted, picked_up, delivered, cancelled]
        riderId:
          type: ["integer", "null"]
        itemsSubtotal:
          type: ["number", "null"]
        deliveryFee:
          type: ["number", "null"]
        totalAmount:
          type: ["number", "null"]
        requestedAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        pickedUpAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
        createdAt:
          type: string
        business:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            slug:
              type: ["string", "null"]
            address:
              type: ["string", "null"]
            latitude:
              type: ["number", "null"]
            longitude:
              type: ["number", "null"]
            phone:
              type: ["string", "null"]
        rider:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            fullName:
              type: ["string", "null"]
            phone:
              type: ["string", "null"]
            vehicleType:
              type: ["string", "null"]
            currentLatitude:
              type: ["number", "null"]
            currentLongitude:
              type: ["number", "null"]
            lastLocationAt:
              type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/DeliveryOrderItem"
        trackingToken:
          description: Guest tracking token, only present once in the response right after creating an order without being logged in.
          type: string
    DeliveryOrderItem:
      type: object
      required: [id, itemName, unitPrice, quantity]
      properties:
        id:
          type: integer
        itemId:
          type: ["integer", "null"]
        itemName:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
...